Make blur distance independent of resolution & FOV
Previously, the image would always be blurred by a certain number of *pixels*, which meant that as the resolution increased, the apparent amount of blur (relative to the world/image) *decreased* (by staying the same as the resolution went up).

This new approach takes the resolution and FOV into account when determining how far across the texture to sample for the blur; this means that a given view of the world will always look (essentially) the same regardless of how "big" it's rendering.

The one problem that this fix reveals, is that for a sufficiently high resolution (or sufficiently narrow FOV), the step between blur samples may be noticeably larger than a single on-screen pixel, which would introduce a visual 'echoing' effect. In this case, the `blur step distance` value may be decreased until the problem is no longer visible -- the only downside to this is that doing so will decrease the distance over which the image is blurred (in world terms), which may be undesired. (This could be compensated for by adding more blur iterations, but as that's hardcoded into the shader as well as the CommandBuffer, there's no simple way to change that without manually editing the shader's code, as far as I'm aware.) Of course, users can also choose to adjust the `blur step distance` for purely aesthetic reasons, to change the range of possible amounts of diffusion that the values in the texture correspond to, so that's just a fun freebie :)